requirements.txt
web3
pandas
requests
joblib
opencv-python
scikit-image
scikit-learn
python-dotenv